# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Short-Term Rental Platform API
  description: |
    Vendor-neutral API for property managers, hosts, and guests.
    Built with privacy-by-design and ethical AI principles.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: Proprietary
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: https://api-staging.example.com/v1
    description: Staging
  - url: http://localhost:4000/v1
    description: Local Development
...
paths:
...
  /messages/threads:
    get:
      tags: [Messages]
      summary: List message threads
      operationId: listThreads
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Message threads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageThread'
  /messages/threads/{threadId}:
    get:
      tags: [Messages]
      summary: Get messages in thread
      operationId: getThread
      security:
        - bearerAuth: []
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Thread messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageThreadDetail'
  /messages/send:
    post:
      tags: [Messages]
      summary: Send message
      operationId: sendMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  # Reviews
...
  /ai/pricing/suggestions:
    get:
      tags: [AI]
      summary: Get AI pricing suggestions
      operationId: getPricingSuggestions
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PropertyId'
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Pricing suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingSuggestions'
  /ai/listing/optimize:
    post:
      tags: [AI]
      summary: Get listing optimization suggestions
      operationId: optimizeListing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptimizeListingRequest'
      responses:
        '200':
          description: Optimization suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingOptimizations'

  # Analytics
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PropertyId:
      name: propertyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        roles:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
    MessageThread:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingId:
          type: string
          format: uuid
        participants:
          type: array
          items:
            $ref: '#/components/schemas/User'
        lastMessage:
          $ref: '#/components/schemas/Message'
        unreadCount:
          type: integer
    MessageThreadDetail:
      type: object
      properties:
        thread:
          $ref: '#/components/schemas/MessageThread'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        senderId:
          type: string
          format: uuid
        content:
          type: string
        sentAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
        isAutoResponse:
          type: boolean
    SendMessageRequest:
      type: object
      required: [recipientId, content]
      properties:
        recipientId:
          type: string
          format: uuid
        bookingId:
          type: string
          format: uuid
        content:
          type: string
          maxLength: 5000

    # Reviews
...
    PricingSuggestions:
      type: object
      properties:
        propertyId:
          type: string
          format: uuid
        suggestions:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              currentPrice:
                type: number
              suggestedPrice:
                type: number
              confidence:
                type: number
              factors:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    impact:
                      type: number
        aiDisclosure:
          type: object
          properties:
            model:
              type: string
            generatedAt:
              type: string
              format: date-time
    OptimizeListingRequest:
      type: object
      required: [propertyId]
      properties:
        propertyId:
          type: string
          format: uuid
        aspects:
          type: array
          items:
            type: string
            enum: [title, description, photos]
    ListingOptimizations:
      type: object
      properties:
        propertyId:
          type: string
          format: uuid
        suggestions:
          type: object
          properties:
            titles:
              type: array
              items:
                type: object
                properties:
                  text:
                    type: string
                  score:
                    type: number
            descriptionImprovements:
              type: array
              items:
                type: object
                properties:
                  section:
                    type: string
                  suggestion:
                    type: string
            photoOrder:
              type: array
              items:
                type: string
                format: uuid
        aiDisclosure:
          type: object
          properties:
            model:
              type: string
            confidence:
              type: number

    # Analytics
...